Unreviewed Chromium debug build fix.
Two fixes makes one breakage

http://trac.webkit.org/changeset/129785
http://trac.webkit.org/changeset/129798

* bindings/v8/V8Binding.h:
(WebCore::toNativeArguments):

Canonical link: https://commits.webkit.org/115818@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129807 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/bindings/v8/V8Binding.h
@@ -224,7 +224,7 @@ namespace WebCore {
     template <class T>
     Vector<T> toNativeArguments(const v8::Arguments& args, int startIndex)
     {
-        ASSERT(startIndex <= static_cast<size_t>(args.Length()));
+        ASSERT(startIndex <= args.Length());
         Vector<T> result;
         typedef NativeValueTraits<T> TraitsType;
         int length = args.Length();
